uploads and downloads complete

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const { validationResult } = require('express-validator');
 const mongoose = require('mongoose');
+const fileHelper = require('../util/file');
 
 const Product = require('../models/product');
 
@@ -10,15 +11,30 @@ exports.getAddProduct = (req, res, next) => {
         editing: false,
         hasError: false,
         errorMessage: null,
-        validationErrors: []
+        validationErrors: [],
     });
 };
 
 exports.postAddProduct = (req, res, next) => {
     const title = req.body.title;
-    const imageUrl = req.file;
+    const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
+    if (!image) {
+        return res.status(422).render('admin/edit-product', {
+            pageTitle: 'Add Product',
+            path: '/admin/add-product',
+            editing: false,
+            hasError: true,
+            product: {
+                title: title,
+                price: price,
+                description: description,
+            },
+            errorMessage: 'Attached file is not an image.',
+            validationErrors: [],
+        });
+    }
     const errors = validationResult(req);
 
     if (!errors.isEmpty()) {
@@ -30,30 +46,31 @@ exports.postAddProduct = (req, res, next) => {
             hasError: true,
             product: {
                 title: title,
-                imageUrl: imageUrl,
                 price: price,
-                description: description
+                description: description,
             },
             errorMessage: errors.array()[0].msg,
-            validationErrors: errors.array()
+            validationErrors: errors.array(),
         });
     }
+
+    const imageUrl = image.path;
 
     const product = new Product({
         title: title,
         price: price,
         description: description,
         imageUrl: imageUrl,
-        userId: req.user
+        userId: req.user,
     });
     product
         .save()
-        .then(result => {
+        .then((result) => {
             // console.log(result);
             console.log('Created Product');
             res.redirect('/admin/products');
         })
-        .catch(err => {
+        .catch((err) => {
             const error = new Error(err);
             error.httpStatusCode = 500;
             return next(error);
@@ -67,7 +84,7 @@ exports.getEditProduct = (req, res, next) => {
     }
     const prodId = req.params.productId;
     Product.findById(prodId)
-        .then(product => {
+        .then((product) => {
             if (!product) {
                 return res.redirect('/');
             }
@@ -78,10 +95,10 @@ exports.getEditProduct = (req, res, next) => {
                 product: product,
                 hasError: false,
                 errorMessage: null,
-                validationErrors: []
+                validationErrors: [],
             });
         })
-        .catch(err => {
+        .catch((err) => {
             const error = new Error(err);
             error.httpStatusCode = 500;
             return next(error);
@@ -92,7 +109,7 @@ exports.postEditProduct = (req, res, next) => {
     const prodId = req.body.productId;
     const updatedTitle = req.body.title;
     const updatedPrice = req.body.price;
-    const updatedImageUrl = req.body.imageUrl;
+    const image = req.file;
     const updatedDesc = req.body.description;
 
     const errors = validationResult(req);
@@ -105,31 +122,33 @@ exports.postEditProduct = (req, res, next) => {
             hasError: true,
             product: {
                 title: updatedTitle,
-                imageUrl: updatedImageUrl,
                 price: updatedPrice,
                 description: updatedDesc,
-                _id: prodId
+                _id: prodId,
             },
             errorMessage: errors.array()[0].msg,
-            validationErrors: errors.array()
+            validationErrors: errors.array(),
         });
     }
 
     Product.findById(prodId)
-        .then(product => {
+        .then((product) => {
             if (product.userId.toString() !== req.user._id.toString()) {
                 return res.redirect('/');
             }
             product.title = updatedTitle;
             product.price = updatedPrice;
             product.description = updatedDesc;
-            product.imageUrl = updatedImageUrl;
-            return product.save().then(result => {
+            if (image) {
+                fileHelper.deleteFile(product.imageUrl);
+                product.imageUrl = image.path;
+            }
+            return product.save().then((result) => {
                 console.log('UPDATED PRODUCT!');
                 res.redirect('/admin/products');
             });
         })
-        .catch(err => {
+        .catch((err) => {
             const error = new Error(err);
             error.httpStatusCode = 500;
             return next(error);
@@ -140,15 +159,15 @@ exports.getProducts = (req, res, next) => {
     Product.find({ userId: req.user._id })
         // .select('title price -_id')
         // .populate('userId', 'name')
-        .then(products => {
+        .then((products) => {
             console.log(products);
             res.render('admin/products', {
                 prods: products,
                 pageTitle: 'Admin Products',
-                path: '/admin/products'
+                path: '/admin/products',
             });
         })
-        .catch(err => {
+        .catch((err) => {
             const error = new Error(err);
             error.httpStatusCode = 500;
             return next(error);
@@ -157,12 +176,19 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({ _id: prodId, userId: req.user._id })
+    Product.findById(prodId)
+        .then((product) => {
+            if (!product) {
+                return next(new Error('Product not found.'));
+            }
+            fileHelper.deleteFile(product.imageUrl);
+            return Product.deleteOne({ _id: prodId, userId: req.user._id });
+        })
         .then(() => {
             console.log('DESTROYED PRODUCT');
             res.redirect('/admin/products');
         })
-        .catch(err => {
+        .catch((err) => {
             const error = new Error(err);
             error.httpStatusCode = 500;
             return next(error);
